+++ src/app/Contracts/BaseRepository.php
@@ -1,6 +1,13 @@
 <?php
 namespace Afrittella\BackProject\Contracts;
 
+/***
+ * Interface BaseRepository
+ * Freely inspired by https://github.com/bosnadev/repository
+ * @package Afrittella\BackProject\Contracts
+ *
+ */
+
 interface BaseRepository
 {
     public function all($columns = array('*'));
@@ -16,4 +23,8 @@ interface BaseRepository
     public function find($id, $columns = array('*'));
 
     public function findBy($field, $value, $columns = array('*'));
+
+    public function findAllBy($field, $value, $columns = array('*'));
+
+    public function findWhere($where, $columns = array('*'), $or = false);
 }